# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/admin/products/{id}:
    put:
      tags:
      - admin-products-rest-controller
      operationId: replaceProduct
      parameters:
      - name: stock_S
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_M
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_L
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_XL
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - admin-products-rest-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/admin/products/{id}/image:
    put:
      tags:
      - admin-products-rest-controller
      operationId: replaceProdcutImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - admin-products-rest-controller
      operationId: createProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-products-rest-controller
      operationId: deleteProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/cart/{id}:
    post:
      tags:
      - cart-rest-controller
      operationId: addToCart
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: size
        in: query
        required: true
        schema:
          type: string
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    patch:
      tags:
      - cart-rest-controller
      operationId: removeFromCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
...
  /api/v1/admin/products:
    get:
      tags:
      - admin-products-rest-controller
      operationId: getAdminProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
    post:
      tags:
      - admin-products-rest-controller
      operationId: createProduct
      parameters:
      - name: stock_S
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_M
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_L
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: stock_XL
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProductRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/cart:
    get:
      tags:
      - cart-rest-controller
      operationId: getCart
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    patch:
      tags:
      - cart-rest-controller
      operationId: proccess
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
...
  /api/v1/products/{id}:
    get:
      tags:
      - product-detail-rest-controller
      operationId: getProductDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/{id}/image:
    get:
      tags:
      - product-detail-rest-controller
      operationId: getProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/admin/products/out-of-stock:
    get:
      tags:
      - admin-products-rest-controller
      operationId: getOutOfStockProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDTO"
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        address:
          type: string
        banned:
          type: boolean
        reports:
          type: integer
          format: int32
        imageUrl:
          type: string
        roles:
          type: array
          items:
            type: string
...
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        user:
          $ref: "#/components/schemas/UserDTO"
        rating:
          type: integer
          format: int32
        reviewText:
          type: string
        reported:
          type: boolean
        own:
          type: boolean
        ratingStars:
          type: array
          items:
            type: boolean
        emptyStars:
          type: array
          items:
            type: boolean
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32
        outOfStock:
          type: boolean
        sold:
          type: integer
          format: int32
        category:
          type: string
        imageBool:
          type: boolean
        imageUrl:
          type: string
        sizes:
          type: array
          items:
            $ref: "#/components/schemas/SizeDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
    SizeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        stock:
          type: integer
          format: int32
...
    BasicProductSizeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32
        outOfStock:
          type: boolean
        sold:
          type: integer
          format: int32
        category:
          type: string
        imageBool:
          type: boolean
        imageUrl:
          type: string
        sizes:
          type: array
          items:
            $ref: "#/components/schemas/SizeDTO"
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        createdAtFormatted:
          type: string
        isPaid:
          type: boolean
        totalPrice:
          type: number
        state:
          type: string
        orderProducts:
          type: array
          items:
            $ref: "#/components/schemas/OrderProductDTO"
    OrderProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: "#/components/schemas/BasicProductSizeDTO"
        productName:
          type: string
        quantity:
          type: integer
          format: int32
        size:
          type: string
        subtotalPrice:
          type: number
          format: double
...
    NewProductRequestDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32
        category:
          type: string
        imageBool:
          type: boolean